Moved MarkdownTagParser into its own file
