+++ Example/Tests/ParserTests.swift
@@ -141,13 +141,13 @@ class ParserTests: XCTestCase {
         do {
             let _ = try MarkdownParser.parse("This _won't__ work because the tags don't match.")
         } catch {
-            XCTAssert(error as! MarkdownTag.ParserError == MarkdownTag.ParserError.tagMismatch)
+            XCTAssert(error as! MarkdownTagParser.ParserError == MarkdownTagParser.ParserError.tagMismatch)
         }
         
         do {
             let _ = try MarkdownParser.parse("Neither **should* this.")
         } catch {
-            XCTAssert(error as! MarkdownTag.ParserError == MarkdownTag.ParserError.tagMismatch)
+            XCTAssert(error as! MarkdownTagParser.ParserError == MarkdownTagParser.ParserError.tagMismatch)
         }
     }
     
@@ -155,13 +155,13 @@ class ParserTests: XCTestCase {
         do {
             let _ = try MarkdownParser.parse("Please *don't _do_ this.")
         } catch {
-            XCTAssert(error as! MarkdownTag.ParserError == MarkdownTag.ParserError.unclosedTags)
+            XCTAssert(error as! MarkdownTagParser.ParserError == MarkdownTagParser.ParserError.unclosedTags)
         }
         
         do {
             let _ = try MarkdownParser.parse("Finish this __sentenc")
         } catch {
-            XCTAssert(error as! MarkdownTag.ParserError == MarkdownTag.ParserError.unclosedTags)
+            XCTAssert(error as! MarkdownTagParser.ParserError == MarkdownTagParser.ParserError.unclosedTags)
         }
     }
